Refactor server code to allow the tests to kill it
I removed the `.listen` call from the main server code to make it more
modular. Now I have an index.js file which imports this code and it does
the job of spinning it up.

This change also allows the tests to spin up an instance of the server
in another port for testing, and kill it afterwards.

While we're at it, I renamed server.js to app.js. I also used `start`
instead of `watch` in the npm scripts.